+++ projects/clr/hipamd/src/hip_platform.hpp
@@ -29,6 +29,8 @@ class PlatformState {
   bool IsValidDynFunc(const void* hfunc);
   hipError_t GetDynFunc(hipFunction_t* hfunc, hipModule_t hmod, const char* func_name);
   hipError_t GetFuncCount(unsigned int* count, hipModule_t hmod);
+  hipError_t EnumerateFunctions(hipFunction_t* functions, unsigned int numFunctions,
+                                hipModule_t hmod);
   hipError_t GetDynGlobalVar(const char* hostVar, hipModule_t hmod, hipDeviceptr_t* dev_ptr,
                              size_t* size_ptr);
   hipError_t GetDynTexRef(const char* hostVar, hipModule_t hmod, textureReference** texRef);