+++ projects/rocprofiler-sdk/source/include/rocprofiler-sdk/hip/runtime_api_id.h
@@ -638,5 +638,8 @@ typedef enum rocprofiler_hip_runtime_api_id_t  // NOLINT(performance-enum-size)
 #if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 33
     ROCPROFILER_HIP_RUNTIME_API_ID_hipInitDevice,
 #endif
+#if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 34
+    ROCPROFILER_HIP_RUNTIME_API_ID_hipModuleEnumerateFunctions,
+#endif
     ROCPROFILER_HIP_RUNTIME_API_ID_LAST,
 } rocprofiler_hip_runtime_api_id_t;